feat: move cache to own file
